requirements.txt
autopep8==1.6.0
Babel==2.9.1
certifi==2021.10.8
Delorean==1.0.0
docopt==0.6.2
getmac==0.8.2
humanize==4.0.0
numpy==1.22.2
pandas==1.4.1
pyasn1==0.4.8
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
rsa==4.7.2
simplejson==3.17.3
six==1.16.0
stomp.py==4.1.24
tigeropen==2.0.7
toml==0.10.2
tzdata==2021.5
tzlocal==4.1
